@@ -1,7 +1,7 @@
 package com.l24o.workcalendar.extensions
 
 import android.content.Context
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 
 
 fun Context.getCompatColor(colorResId: Int): Int {
