@@ -1,6 +1,8 @@
 package com.l24o.workcalendar.extensions
 
+import org.threeten.bp.LocalDate
 import java.text.SimpleDateFormat
+import java.time.ZoneId
 import java.util.*
 
 /**
@@ -11,32 +13,10 @@ fun Date.toString(format: String): String {
     return SimpleDateFormat(format, Locale.getDefault()).format(this)
 }
 
-fun Date.isSameMouth(month: Int): Boolean {
-    val calendar = Calendar.getInstance()
-    calendar.time = this
-    return month == calendar.get(Calendar.MONTH)
+fun LocalDate.isSameDay(date: LocalDate): Boolean {
+    return this.dayOfYear == date.dayOfYear
 }
 
-fun Date.isSameDay(date: Date): Boolean {
-    val cal1 = Calendar.getInstance()
-    val cal2 = Calendar.getInstance()
-    cal1.time = this
-    cal2.time = date
-    return cal1.get(Calendar.DAY_OF_YEAR) === cal2.get(Calendar.DAY_OF_YEAR)
-}
-
-fun Date.monthNumber(): Int {
-    val cal1 = Calendar.getInstance()
-    cal1.time = this
-    return cal1.get(Calendar.MONTH)
-}
-
-fun Date.startOfDay(): Date {
-    val cal = Calendar.getInstance()
-    cal.time = this
-    cal.set(Calendar.HOUR_OF_DAY, cal.getMinimum(Calendar.HOUR_OF_DAY));
-    cal.set(Calendar.MINUTE, cal.getMinimum(Calendar.MINUTE));
-    cal.set(Calendar.SECOND, cal.getMinimum(Calendar.SECOND));
-    cal.set(Calendar.MILLISECOND, cal.getMinimum(Calendar.MILLISECOND));
-    return cal.time
+fun LocalDate.isToday(): Boolean {
+    return this.dayOfYear == LocalDate.now().dayOfYear
 }
