@@ -4,6 +4,7 @@ import org.simpleframework.xml.Attribute
 import org.simpleframework.xml.Element
 import org.simpleframework.xml.ElementList
 import org.simpleframework.xml.Root
+import org.threeten.bp.LocalDate
 import java.util.*
 
 /**
@@ -12,9 +13,9 @@ import java.util.*
 @Root(name = "calendar")
 data class Calendar(
         @field:ElementList(entry = "holidays")
-        var holidays: List<Holiday>? = null,
+        var holidays: List<Holiday> = mutableListOf(),
         @field:ElementList(name = "days")
-        var days: List<Day>? = null,
+        var days: List<Day> = mutableListOf(),
         @field:Attribute(name = "year")
         var year: String? = null,
         @field:Attribute(name = "lang")
@@ -26,19 +27,23 @@ data class Calendar(
 @Element(name = "holiday")
 data class Holiday(
         @field:Attribute(name = "id")
-        var id: String? = null,
+        var id: String = "",
         @field:Attribute(name = "title")
-        var title: String? = null,
-        var date: ArrayList<Date> = ArrayList()
-)
+        var title: String = "",
+        var date: ArrayList<LocalDate> = ArrayList()
+) {
+    override fun equals(other: Any?): Boolean {
+        return other is Holiday && id == other.id
+    }
+}
 
 @Element(name = "day")
 data class Day(
         @field:Attribute(name = "d")
-        var day: Date? = null,
+        var day: LocalDate = LocalDate.now(),
         @field:Attribute(name = "t")
-        var type: TypeOfDay? = null,
-        @field:Attribute(name = "h", empty = "", required = false)
+        var type: TypeOfDay = TypeOfDay.NONE,
+        @field:Attribute(name = "h", required = false)
         var holidayId: String? = null
 )
 
